pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "terminaleyes"
version = "0.1.0"
description = "Vision-based agentic terminal controller using webcam input and keystroke output"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "PyYAML>=6.0",
    "pygame>=2.5.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "mypy>=1.8.0",
    "ruff>=0.5.0",
    "pre-commit>=3.6.0",
]
rpi = [
    # Future: Raspberry Pi USB HID dependencies
    # "pyserial>=3.5",
]

[project.scripts]
terminaleyes = "terminaleyes.cli:main"
terminaleyes-endpoint = "terminaleyes.endpoint.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/terminaleyes"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow running",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "SIM", "TCH"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.lint.isort]
known-first-party = ["terminaleyes"]
